[ICU-900] Fixed some compiler warnings. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/8696 Former-commit-id: d63cb9aa9cf0af4eab44d4b94b4f0f1fa155de29

+++ icu4c/source/test/cintltst/citertst.c
@@ -1473,7 +1473,7 @@ static void TestDiscontiguos() {
             if (e == 0) {
                 e = u_strchr(s, 0);
             }
-            ucol_setText(resultiter, s, e - s, &status);
+            ucol_setText(resultiter, s, (int32_t)(e - s), &status);
             ce = ucol_next(resultiter, &status);
             if (U_FAILURE(status)) {
                 log_err("Error manipulating collation iterator\n");
